Use int for fileDescriptionPacket file byte count

+++ cmd/par/main.go
@@ -87,7 +87,7 @@ func (par2LogDecoderDelegate) OnMainPacketLoad(sliceByteCount, recoverySetCount,
 	fmt.Printf("Loaded main packet: slice byte count=%d, recovery set size=%d, non-recovery set size=%d\n", sliceByteCount, recoverySetCount, nonRecoverySetCount)
 }
 
-func (par2LogDecoderDelegate) OnFileDescriptionPacketLoad(fileID [16]byte, filename string, byteCount uint64) {
+func (par2LogDecoderDelegate) OnFileDescriptionPacketLoad(fileID [16]byte, filename string, byteCount int) {
 	fmt.Printf("Loaded file description packet for %q (ID=%x, %d bytes)\n", filename, fileID, byteCount)
 }
